// client/src/app/components/svg-attributes.ts
export enum SvgTypes {
    TSPAN = 'tspan',
    TEXT = 'text',
    SVG_LINK = 'http://www.w3.org/2000/svg',
    SVG = 'svg',
    RECT = 'rect',
    POLYGON = 'polygon',
    ELLIPSE = 'ellipse',
    PATH = 'path',
    IMAGE = 'image',
    CIRCLE = 'circle',
    A = 'a',
    X = 'x',
    G = 'g',
    DIV = 'DIV',
}

export enum SvgAttributes {
    FILL = 'fill',
    STROKE = 'stroke',
    STROKE_WIDTH = 'stroke-width',
    HEIGHT = 'height',
    WIDTH = 'width',
    X = 'x',
    Y = 'y',
    STROKE_DASHARRAY = 'stroke-dasharray',
    POINTS = 'points',
    RX = 'rx',
    RY = 'ry',
    CX = 'cx',
    CY = 'cy',
    R = 'r',
    D = 'd',
    STROKE_LINECAP = 'stroke-linecap',
    L = 'L',
    M = 'M',
    TRANSFORM = 'transform',
    ROTATE = 'rotate',
    HREF = 'href',
    FILTER = 'filter',
    URL = 'url',
    STROKE_LINEJOIN = 'stroke-linejoin',
    STYLE = 'style',
    DY = 'dy',
    FONT_SIZE = 'font-size',
    FONT_STYLE = 'font-style',
    FONT_WEIGHT = 'font-weight',
    TEXT_ANCHOR = 'text-anchor',
    DOWNLOAD = 'download',
    ID= 'id',
    FILL_OPACITY= 'fill-opacity',
    TRANSLATE = 'translate',
    OPACITY= 'opacity',
    SCALE= 'scale',
    Z = 'Z',
    FILL_RULE = 'fill-rule',

}